refactor: updates on paths
