+++ headlesscms-angular/src/app/blogs/blogs.component.html
@@ -32,7 +32,7 @@
               Url: {{blog.url}}
             </mat-card-actions>
             <mat-card-footer>
-              Hashtags: {{blog.topic}}
+              Topic: {{blog.topic}}
             </mat-card-footer>
           </mat-card>
         </mat-card-content>
